+++ packages/observable/src/operators/index.ts
@@ -1,0 +1,2 @@
+export * from './filter-action';
+export * from './map-action';